feat(extract/nvd/feed/cve/v2): add Strategy E selective cpematch expansion
SEMVER ranges now expand cpematch and emit exact-match criteria for the
versions the range criterion cannot cover — non-semver versions plus the
rare semver version NVD lists outside the declared range. Unknown ranges
keep expanding every version.

Switch cpematch resolution to a matchCriteriaId UUID directory index
(removes the FNV-hash + productVariants fallback that worked around the
nx-os/nx_os spelling mismatch between the CVE and cpematch feeds).

Relocate pkg/extract/nvd/feed/v2 → pkg/extract/nvd/feed/cve/v2 to follow
the feed/<datatype>/<version> convention; the orphaned old feed/cve/v2
extractor (no CLI wiring) is removed.

Add the semver-range-mixed golden fixture (CVE-2021-1569 cisco:jabber)
covering both the skip and add paths.
