+++ pkg/extract/nvd/feed/cve/v2/cpe.go
