--- pkg/extract/nvd/feed/cve/v2/testdata/fixtures/2021/CVE-2021-23814.json
@@ -1,165 +1,0 @@
-{
-	"id": "CVE-2021-23814",
-	"sourceIdentifier": "report@snyk.io",
-	"vulnStatus": "Modified",
-	"published": "2021-12-17T20:15:08.340",
-	"lastModified": "2025-06-17T12:15:20.773",
-	"descriptions": [
-		{
-			"lang": "en",
-			"value": "This affects versions of the package unisharp/laravel-filemanager before 2.6.2. The upload() function does not sufficiently validate the file type when uploading.\r\rAn attacker may be able to reproduce the following steps:\r\r1. Install a package with a web Laravel application.\r2. Navigate to the Upload window\r3. Upload an image file, then capture the request\r4. Edit the request contents with a malicious file (webshell)\r5. Enter the path of file uploaded on URL - Remote Code Execution\r\r\r**Note:** Prevention for bad extensions can be done by using a whitelist in the config file(lfm.php). Corresponding document can be found in [here](https://unisharp.github.io/laravel-filemanager/configfolder-categories)."
-		},
-		{
-			"lang": "es",
-			"value": "Esto afecta al paquete unisharp/laravel-filemanager desde la versión 0.0.0. La función upload() no comprueba suficientemente el tipo de archivo cuando es cargado. Un atacante puede ser capaz de reproducir los siguientes pasos: - Instalar un paquete con una aplicación web Laravel. - Navegar a la ventana de subida - Subir un archivo de imagen, luego capturar la petición - Editar el contenido de la petición con un archivo malicioso (webshell) - Introducir la ruta del archivo subido en la URL - Ejecución de Código Remota. **Nota: La prevención para extensiones malas puede hacerse usando una lista blanca en el archivo de configuración (lfm.php). El documento correspondiente puede encontrarse en [aquí](https://unisharp.github.io/laravel-filemanager/configfolder-categories)"
-		}
-	],
-	"references": [
-		{
-			"source": "report@snyk.io",
-			"tags": [
-				"Broken Link",
-				"Third Party Advisory"
-			],
-			"url": "https://github.com/UniSharp/laravel-filemanager/blob/master/src/Controllers/UploadController.php%23L26"
-		},
-		{
-			"source": "report@snyk.io",
-			"url": "https://github.com/UniSharp/laravel-filemanager/issues/1113"
-		},
-		{
-			"source": "report@snyk.io",
-			"url": "https://github.com/UniSharp/laravel-filemanager/releases/tag/v2.5.1"
-		},
-		{
-			"source": "report@snyk.io",
-			"url": "https://github.com/UniSharp/laravel-filemanager/releases/tag/v2.6.2"
-		},
-		{
-			"source": "report@snyk.io",
-			"url": "https://security.snyk.io/vuln/SNYK-PHP-UNISHARPLARAVELFILEMANAGER-1567199"
-		},
-		{
-			"source": "af854a3a-2127-422b-91ae-364da2661108",
-			"tags": [
-				"Broken Link",
-				"Third Party Advisory"
-			],
-			"url": "https://github.com/UniSharp/laravel-filemanager/blob/master/src/Controllers/UploadController.php%23L26"
-		},
-		{
-			"source": "af854a3a-2127-422b-91ae-364da2661108",
-			"url": "https://github.com/UniSharp/laravel-filemanager/issues/1113#issuecomment-1812092975"
-		},
-		{
-			"source": "af854a3a-2127-422b-91ae-364da2661108",
-			"tags": [
-				"Third Party Advisory"
-			],
-			"url": "https://snyk.io/vuln/SNYK-PHP-UNISHARPLARAVELFILEMANAGER-1567199"
-		}
-	],
-	"metrics": {
-		"cvssMetricV2": [
-			{
-				"source": "nvd@nist.gov",
-				"type": "Primary",
-				"cvssData": {
-					"version": "2.0",
-					"vectorString": "AV:N/AC:L/Au:S/C:P/I:P/A:P",
-					"accessVector": "NETWORK",
-					"accessComplexity": "LOW",
-					"authentication": "SINGLE",
-					"confidentialityImpact": "PARTIAL",
-					"integrityImpact": "PARTIAL",
-					"availabilityImpact": "PARTIAL",
-					"baseScore": 6.5
-				},
-				"baseSeverity": "MEDIUM",
-				"exploitabilityScore": 8,
-				"impactScore": 6.4
-			}
-		],
-		"cvssMetricV31": [
-			{
-				"source": "report@snyk.io",
-				"type": "Secondary",
-				"cvssData": {
-					"version": "3.1",
-					"vectorString": "CVSS:3.1/AV:N/AC:H/PR:L/UI:R/S:U/C:H/I:H/A:L",
-					"attackVector": "NETWORK",
-					"attackComplexity": "HIGH",
-					"privilegesRequired": "LOW",
-					"userInteraction": "REQUIRED",
-					"scope": "UNCHANGED",
-					"confidentialityImpact": "HIGH",
-					"integrityImpact": "HIGH",
-					"availabilityImpact": "LOW",
-					"baseScore": 6.7,
-					"baseSeverity": "MEDIUM"
-				},
-				"exploitabilityScore": 1.2,
-				"impactScore": 5.5
-			},
-			{
-				"source": "nvd@nist.gov",
-				"type": "Primary",
-				"cvssData": {
-					"version": "3.1",
-					"vectorString": "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:H/A:H",
-					"attackVector": "NETWORK",
-					"attackComplexity": "LOW",
-					"privilegesRequired": "LOW",
-					"userInteraction": "NONE",
-					"scope": "UNCHANGED",
-					"confidentialityImpact": "HIGH",
-					"integrityImpact": "HIGH",
-					"availabilityImpact": "HIGH",
-					"baseScore": 8.8,
-					"baseSeverity": "HIGH"
-				},
-				"exploitabilityScore": 2.8,
-				"impactScore": 5.9
-			}
-		]
-	},
-	"weaknesses": [
-		{
-			"source": "report@snyk.io",
-			"type": "Secondary",
-			"description": [
-				{
-					"lang": "en",
-					"value": "CWE-94"
-				}
-			]
-		},
-		{
-			"source": "nvd@nist.gov",
-			"type": "Primary",
-			"description": [
-				{
-					"lang": "en",
-					"value": "CWE-434"
-				}
-			]
-		}
-	],
-	"configurations": [
-		{
-			"nodes": [
-				{
-					"operator": "OR",
-					"cpeMatch": [
-						{
-							"vulnerable": true,
-							"criteria": "cpe:2.3:a:unisharp:laravel-filemanager:*:*:*:*:*:*:*:*",
-							"matchCriteriaId": "4467027F-A375-40AE-AAC5-2A6CD3BC9598",
-							"versionStartIncluding": "0.0.0"
-						}
-					]
-				}
-			]
-		}
-	]
-}
