--- pkg/extract/nvd/feed/cve/v2/testdata/fixtures/2021/CVE-2021-3002.json
@@ -1,126 +1,0 @@
-{
-	"id": "CVE-2021-3002",
-	"sourceIdentifier": "cve@mitre.org",
-	"vulnStatus": "Modified",
-	"published": "2021-01-01T19:15:11.077",
-	"lastModified": "2024-11-21T06:20:43.243",
-	"descriptions": [
-		{
-			"lang": "en",
-			"value": "Seo Panel 4.8.0 allows reflected XSS via the seo/seopanel/login.php?sec=forgot email parameter."
-		},
-		{
-			"lang": "es",
-			"value": "Seo Panel versión 4.8.0, permite un ataque de tipo XSS reflejado por medio del parámetro email de seo/seopanel/login.php?sec=forgot."
-		}
-	],
-	"references": [
-		{
-			"source": "cve@mitre.org",
-			"tags": [
-				"Exploit",
-				"Third Party Advisory"
-			],
-			"url": "http://www.cinquino.eu/SeoPanelReflect.htm"
-		},
-		{
-			"source": "cve@mitre.org",
-			"tags": [
-				"Exploit",
-				"Issue Tracking",
-				"Third Party Advisory"
-			],
-			"url": "https://github.com/seopanel/Seo-Panel/issues/202"
-		},
-		{
-			"source": "af854a3a-2127-422b-91ae-364da2661108",
-			"tags": [
-				"Exploit",
-				"Third Party Advisory"
-			],
-			"url": "http://www.cinquino.eu/SeoPanelReflect.htm"
-		},
-		{
-			"source": "af854a3a-2127-422b-91ae-364da2661108",
-			"tags": [
-				"Exploit",
-				"Issue Tracking",
-				"Third Party Advisory"
-			],
-			"url": "https://github.com/seopanel/Seo-Panel/issues/202"
-		}
-	],
-	"metrics": {
-		"cvssMetricV2": [
-			{
-				"source": "nvd@nist.gov",
-				"type": "Primary",
-				"cvssData": {
-					"version": "2.0",
-					"vectorString": "AV:N/AC:M/Au:N/C:N/I:P/A:N",
-					"accessVector": "NETWORK",
-					"accessComplexity": "MEDIUM",
-					"authentication": "NONE",
-					"confidentialityImpact": "NONE",
-					"integrityImpact": "PARTIAL",
-					"availabilityImpact": "NONE",
-					"baseScore": 4.3
-				},
-				"baseSeverity": "MEDIUM",
-				"exploitabilityScore": 8.6,
-				"impactScore": 2.9,
-				"userInteractionRequired": true
-			}
-		],
-		"cvssMetricV31": [
-			{
-				"source": "nvd@nist.gov",
-				"type": "Primary",
-				"cvssData": {
-					"version": "3.1",
-					"vectorString": "CVSS:3.1/AV:N/AC:L/PR:N/UI:R/S:C/C:L/I:L/A:N",
-					"attackVector": "NETWORK",
-					"attackComplexity": "LOW",
-					"privilegesRequired": "NONE",
-					"userInteraction": "REQUIRED",
-					"scope": "CHANGED",
-					"confidentialityImpact": "LOW",
-					"integrityImpact": "LOW",
-					"availabilityImpact": "NONE",
-					"baseScore": 6.1,
-					"baseSeverity": "MEDIUM"
-				},
-				"exploitabilityScore": 2.8,
-				"impactScore": 2.7
-			}
-		]
-	},
-	"weaknesses": [
-		{
-			"source": "nvd@nist.gov",
-			"type": "Primary",
-			"description": [
-				{
-					"lang": "en",
-					"value": "CWE-79"
-				}
-			]
-		}
-	],
-	"configurations": [
-		{
-			"nodes": [
-				{
-					"operator": "OR",
-					"cpeMatch": [
-						{
-							"vulnerable": true,
-							"criteria": "cpe:2.3:a:seopanel:seo_panel:4.8.0:*:*:*:*:*:*:*",
-							"matchCriteriaId": "FA338EF9-AC8B-411D-8BE3-9F9A2E68F8E0"
-						}
-					]
-				}
-			]
-		}
-	]
-}
