+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/exact-match/vuls-data-raw-nvd-feed-cpematch-v2/67f7a4ef/2D49E611-5D53-479D-A981-42388FDC0E8D.json
