+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/exact-match/vuls-data-raw-nvd-feed-cve-v2/2024/CVE-2024-0028.json
