+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/non-semver-range/vuls-data-raw-nvd-feed-cve-v2/2024/CVE-2024-20282.json
