+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/semver-range-mixed/vuls-data-raw-nvd-feed-cpematch-v2/81835afe/32CADB26-D0A9-46AD-86CD-445AA8285AA5.json
@@ -1,0 +1,16 @@
+{
+	"criteria": "cpe:2.3:a:cisco:jabber:*:*:*:*:*:windows:*:*",
+	"matchCriteriaId": "32CADB26-D0A9-46AD-86CD-445AA8285AA5",
+	"versionStartIncluding": "14.0",
+	"versionEndExcluding": "14.0.1.55914",
+	"created": "2021-06-23T14:39:49.270",
+	"lastModified": "2021-06-23T14:39:49.270",
+	"cpeLastModified": "2021-06-23T14:40:15.207",
+	"status": "Active",
+	"matches": [
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:14.0:*:*:*:*:windows:*:*",
+			"cpeNameId": "E4438E64-EC0D-4B62-B167-9544B4B092D0"
+		}
+	]
+}
