+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/semver-range-mixed/vuls-data-raw-nvd-feed-cpematch-v2/81835afe/A1E7CD95-CBC7-4876-8F03-D719AAD9EE06.json
@@ -1,0 +1,44 @@
+{
+	"criteria": "cpe:2.3:a:cisco:jabber:*:*:*:*:*:windows:*:*",
+	"matchCriteriaId": "A1E7CD95-CBC7-4876-8F03-D719AAD9EE06",
+	"versionStartIncluding": "12.9",
+	"versionEndExcluding": "12.9.6.55898",
+	"created": "2021-06-23T14:39:49.270",
+	"lastModified": "2021-06-23T14:39:49.270",
+	"cpeLastModified": "2021-06-23T14:39:49.270",
+	"status": "Active",
+	"matches": [
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9:*:*:*:*:windows:*:*",
+			"cpeNameId": "0A8A8A70-0399-449A-96C1-18FC44EA8E0E"
+		},
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9\\(0\\):*:*:*:*:windows:*:*",
+			"cpeNameId": "08BD7B89-7DAC-48A2-99BC-0FB0C1141995"
+		},
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9\\(1\\):*:*:*:*:windows:*:*",
+			"cpeNameId": "89258A36-7BC4-4AB3-93A4-B7494975148D"
+		},
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9\\(2\\):*:*:*:*:windows:*:*",
+			"cpeNameId": "62ECEDD1-2E8C-4D15-B4B2-FCEEEDE587A6"
+		},
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9\\(3\\):*:*:*:*:windows:*:*",
+			"cpeNameId": "0C385E1C-197D-4429-A06C-D32C63FCF576"
+		},
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9\\(4\\):*:*:*:*:windows:*:*",
+			"cpeNameId": "F2B3C545-A858-4B9A-AD54-FFCB74D9E67D"
+		},
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9.1:*:*:*:*:windows:*:*",
+			"cpeNameId": "D1F1B23D-EE05-4FE6-8E1B-0615F4424566"
+		},
+		{
+			"cpeName": "cpe:2.3:a:cisco:jabber:12.9.3:*:*:*:*:windows:*:*",
+			"cpeNameId": "903A7FE4-1E0C-472C-9826-162D1902847C"
+		}
+	]
+}
