+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/semver-range-mixed/vuls-data-raw-nvd-feed-cve-v2/2021/CVE-2021-1569.json
@@ -1,0 +1,157 @@
+{
+	"id": "CVE-2021-1569",
+	"sourceIdentifier": "psirt@cisco.com",
+	"vulnStatus": "Modified",
+	"published": "2021-06-16T18:15:09.077",
+	"lastModified": "2024-11-21T05:44:38.637",
+	"descriptions": [
+		{
+			"lang": "en",
+			"value": "Multiple vulnerabilities in Cisco Jabber for Windows, Cisco Jabber for Mac, and Cisco Jabber for mobile platforms could allow an attacker to access sensitive information or cause a denial of service (DoS) condition. For more information about these vulnerabilities, see the Details section of this advisory."
+		},
+		{
+			"lang": "es",
+			"value": "Múltiples vulnerabilidades en Cisco Jabber for Windows, Cisco Jabber for Mac y Cisco Jabber para plataformas móviles, podrían permitir a un atacante acceder a información confidencial o causar una condición de denegación de servicio (DoS). Para mayor información sobre estas vulnerabilidades, consulte la sección Detalles de este aviso"
+		}
+	],
+	"references": [
+		{
+			"source": "psirt@cisco.com",
+			"tags": [
+				"Vendor Advisory"
+			],
+			"url": "https://tools.cisco.com/security/center/content/CiscoSecurityAdvisory/cisco-sa-jabber-GuC5mLwG"
+		},
+		{
+			"source": "af854a3a-2127-422b-91ae-364da2661108",
+			"tags": [
+				"Vendor Advisory"
+			],
+			"url": "https://tools.cisco.com/security/center/content/CiscoSecurityAdvisory/cisco-sa-jabber-GuC5mLwG"
+		}
+	],
+	"metrics": {
+		"cvssMetricV2": [
+			{
+				"source": "nvd@nist.gov",
+				"type": "Primary",
+				"cvssData": {
+					"version": "2.0",
+					"vectorString": "AV:N/AC:L/Au:S/C:P/I:N/A:N",
+					"accessVector": "NETWORK",
+					"accessComplexity": "LOW",
+					"authentication": "SINGLE",
+					"confidentialityImpact": "PARTIAL",
+					"integrityImpact": "NONE",
+					"availabilityImpact": "NONE",
+					"baseScore": 4,
+					"temporalScore": 0,
+					"environmentalScore": 0
+				},
+				"baseSeverity": "MEDIUM",
+				"exploitabilityScore": 8,
+				"impactScore": 2.9,
+				"acInsufInfo": false,
+				"obtainAllPrivilege": false,
+				"obtainUserPrivilege": false,
+				"obtainOtherPrivilege": false,
+				"userInteractionRequired": false
+			}
+		],
+		"cvssMetricV31": [
+			{
+				"source": "psirt@cisco.com",
+				"type": "Secondary",
+				"cvssData": {
+					"version": "3.1",
+					"vectorString": "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:N/A:N",
+					"attackVector": "NETWORK",
+					"attackComplexity": "LOW",
+					"privilegesRequired": "LOW",
+					"userInteraction": "NONE",
+					"scope": "UNCHANGED",
+					"confidentialityImpact": "HIGH",
+					"integrityImpact": "NONE",
+					"availabilityImpact": "NONE",
+					"baseScore": 6.5,
+					"baseSeverity": "MEDIUM",
+					"temporalScore": 0,
+					"environmentalScore": 0
+				},
+				"exploitabilityScore": 2.8,
+				"impactScore": 3.6
+			},
+			{
+				"source": "nvd@nist.gov",
+				"type": "Primary",
+				"cvssData": {
+					"version": "3.1",
+					"vectorString": "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:N/A:N",
+					"attackVector": "NETWORK",
+					"attackComplexity": "LOW",
+					"privilegesRequired": "LOW",
+					"userInteraction": "NONE",
+					"scope": "UNCHANGED",
+					"confidentialityImpact": "HIGH",
+					"integrityImpact": "NONE",
+					"availabilityImpact": "NONE",
+					"baseScore": 6.5,
+					"baseSeverity": "MEDIUM",
+					"temporalScore": 0,
+					"environmentalScore": 0
+				},
+				"exploitabilityScore": 2.8,
+				"impactScore": 3.6
+			}
+		]
+	},
+	"weaknesses": [
+		{
+			"source": "psirt@cisco.com",
+			"type": "Secondary",
+			"description": [
+				{
+					"lang": "en",
+					"value": "CWE-399"
+				}
+			]
+		},
+		{
+			"source": "nvd@nist.gov",
+			"type": "Primary",
+			"description": [
+				{
+					"lang": "en",
+					"value": "CWE-20"
+				}
+			]
+		}
+	],
+	"configurations": [
+		{
+			"negate": false,
+			"nodes": [
+				{
+					"operator": "OR",
+					"negate": false,
+					"cpeMatch": [
+						{
+							"vulnerable": true,
+							"criteria": "cpe:2.3:a:cisco:jabber:*:*:*:*:*:windows:*:*",
+							"matchCriteriaId": "A1E7CD95-CBC7-4876-8F03-D719AAD9EE06",
+							"versionStartIncluding": "12.9",
+							"versionEndExcluding": "12.9.6.55898"
+						},
+						{
+							"vulnerable": true,
+							"criteria": "cpe:2.3:a:cisco:jabber:*:*:*:*:*:windows:*:*",
+							"matchCriteriaId": "32CADB26-D0A9-46AD-86CD-445AA8285AA5",
+							"versionStartIncluding": "14.0",
+							"versionEndExcluding": "14.0.1.55914"
+						}
+					]
+				}
+			]
+		}
+	]
+}
