+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/semver-range/vuls-data-raw-nvd-feed-cpematch-v2/ce4c4fae/E280D93F-B51F-4F59-9CCE-829C1F9F1A78.json
