+++ pkg/extract/nvd/feed/cve/v2/testdata/fixtures/semver-range/vuls-data-raw-nvd-feed-cve-v2/2024/CVE-2024-0008.json
