+++ pkg/extract/nvd/feed/cve/v2/testdata/golden/exact-match/data/2024/CVE-2024-0028.json
