+++ pkg/extract/nvd/feed/cve/v2/testdata/golden/exact-match/datasource.json
