+++ pkg/extract/nvd/feed/cve/v2/testdata/golden/non-semver-range/data/2024/CVE-2024-20282.json
