+++ pkg/extract/nvd/feed/cve/v2/testdata/golden/semver-range-mixed/data/2021/CVE-2021-1569.json
@@ -1,0 +1,238 @@
+{
+	"id": "CVE-2021-1569",
+	"vulnerabilities": [
+		{
+			"content": {
+				"id": "CVE-2021-1569",
+				"description": "Multiple vulnerabilities in Cisco Jabber for Windows, Cisco Jabber for Mac, and Cisco Jabber for mobile platforms could allow an attacker to access sensitive information or cause a denial of service (DoS) condition. For more information about these vulnerabilities, see the Details section of this advisory.",
+				"severity": [
+					{
+						"type": "cvss_v2",
+						"source": "nvd@nist.gov",
+						"cvss_v2": {
+							"vector": "AV:N/AC:L/Au:S/C:P/I:N/A:N",
+							"base_score": 4,
+							"nvd_base_severity": "MEDIUM",
+							"temporal_score": 4,
+							"nvd_temporal_severity": "MEDIUM",
+							"environmental_score": 4,
+							"nvd_environmental_severity": "MEDIUM"
+						}
+					},
+					{
+						"type": "cvss_v31",
+						"source": "nvd@nist.gov",
+						"cvss_v31": {
+							"vector": "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:N/A:N",
+							"base_score": 6.5,
+							"base_severity": "MEDIUM",
+							"temporal_score": 6.5,
+							"temporal_severity": "MEDIUM",
+							"environmental_score": 6.5,
+							"environmental_severity": "MEDIUM"
+						}
+					},
+					{
+						"type": "cvss_v31",
+						"source": "psirt@cisco.com",
+						"cvss_v31": {
+							"vector": "CVSS:3.1/AV:N/AC:L/PR:L/UI:N/S:U/C:H/I:N/A:N",
+							"base_score": 6.5,
+							"base_severity": "MEDIUM",
+							"temporal_score": 6.5,
+							"temporal_severity": "MEDIUM",
+							"environmental_score": 6.5,
+							"environmental_severity": "MEDIUM"
+						}
+					}
+				],
+				"cwe": [
+					{
+						"source": "nvd@nist.gov",
+						"cwe": [
+							"CWE-20"
+						]
+					},
+					{
+						"source": "psirt@cisco.com",
+						"cwe": [
+							"CWE-399"
+						]
+					}
+				],
+				"references": [
+					{
+						"source": "af854a3a-2127-422b-91ae-364da2661108",
+						"url": "https://tools.cisco.com/security/center/content/CiscoSecurityAdvisory/cisco-sa-jabber-GuC5mLwG"
+					},
+					{
+						"source": "nvd.nist.gov",
+						"url": "https://nvd.nist.gov/vuln/detail/CVE-2021-1569"
+					},
+					{
+						"source": "psirt@cisco.com",
+						"url": "https://tools.cisco.com/security/center/content/CiscoSecurityAdvisory/cisco-sa-jabber-GuC5mLwG"
+					}
+				],
+				"published": "2021-06-16T18:15:09.077Z",
+				"modified": "2024-11-21T05:44:38.637Z"
+			},
+			"segments": [
+				{
+					"ecosystem": "cpe"
+				}
+			]
+		}
+	],
+	"detections": [
+		{
+			"ecosystem": "cpe",
+			"conditions": [
+				{
+					"criteria": {
+						"operator": "OR",
+						"criterias": [
+							{
+								"operator": "OR",
+								"criterias": [
+									{
+										"operator": "OR",
+										"criterias": [
+											{
+												"operator": "OR",
+												"criterions": [
+													{
+														"type": "version",
+														"version": {
+															"vulnerable": true,
+															"fix_status": {
+																"class": "unknown"
+															},
+															"package": {
+																"type": "cpe",
+																"cpe": "cpe:2.3:a:cisco:jabber:*:*:*:*:*:windows:*:*"
+															},
+															"affected": {
+																"type": "semver",
+																"range": [
+																	{
+																		"lt": "12.9.6.55898",
+																		"ge": "12.9"
+																	}
+																]
+															}
+														}
+													},
+													{
+														"type": "version",
+														"version": {
+															"vulnerable": true,
+															"fix_status": {
+																"class": "unknown"
+															},
+															"package": {
+																"type": "cpe",
+																"cpe": "cpe:2.3:a:cisco:jabber:12.9\\(0\\):*:*:*:*:windows:*:*"
+															}
+														}
+													},
+													{
+														"type": "version",
+														"version": {
+															"vulnerable": true,
+															"fix_status": {
+																"class": "unknown"
+															},
+															"package": {
+																"type": "cpe",
+																"cpe": "cpe:2.3:a:cisco:jabber:12.9\\(1\\):*:*:*:*:windows:*:*"
+															}
+														}
+													},
+													{
+														"type": "version",
+														"version": {
+															"vulnerable": true,
+															"fix_status": {
+																"class": "unknown"
+															},
+															"package": {
+																"type": "cpe",
+																"cpe": "cpe:2.3:a:cisco:jabber:12.9\\(2\\):*:*:*:*:windows:*:*"
+															}
+														}
+													},
+													{
+														"type": "version",
+														"version": {
+															"vulnerable": true,
+															"fix_status": {
+																"class": "unknown"
+															},
+															"package": {
+																"type": "cpe",
+																"cpe": "cpe:2.3:a:cisco:jabber:12.9\\(3\\):*:*:*:*:windows:*:*"
+															}
+														}
+													},
+													{
+														"type": "version",
+														"version": {
+															"vulnerable": true,
+															"fix_status": {
+																"class": "unknown"
+															},
+															"package": {
+																"type": "cpe",
+																"cpe": "cpe:2.3:a:cisco:jabber:12.9\\(4\\):*:*:*:*:windows:*:*"
+															}
+														}
+													}
+												]
+											},
+											{
+												"operator": "OR",
+												"criterions": [
+													{
+														"type": "version",
+														"version": {
+															"vulnerable": true,
+															"fix_status": {
+																"class": "unknown"
+															},
+															"package": {
+																"type": "cpe",
+																"cpe": "cpe:2.3:a:cisco:jabber:*:*:*:*:*:windows:*:*"
+															},
+															"affected": {
+																"type": "semver",
+																"range": [
+																	{
+																		"lt": "14.0.1.55914",
+																		"ge": "14.0"
+																	}
+																]
+															}
+														}
+													}
+												]
+											}
+										]
+									}
+								]
+							}
+						]
+					}
+				}
+			]
+		}
+	],
+	"data_source": {
+		"id": "nvd-feed-cve-v2",
+		"raws": [
+			"vuls-data-raw-nvd-feed-cpematch-v2/81835afe/32CADB26-D0A9-46AD-86CD-445AA8285AA5.json",
+			"vuls-data-raw-nvd-feed-cpematch-v2/81835afe/A1E7CD95-CBC7-4876-8F03-D719AAD9EE06.json",
+			"vuls-data-raw-nvd-feed-cve-v2/2021/CVE-2021-1569.json"
+		]
+	}
+}
