+++ pkg/extract/nvd/feed/cve/v2/testdata/golden/semver-range-mixed/datasource.json
