+++ pkg/extract/nvd/feed/cve/v2/testdata/golden/semver-range/data/2024/CVE-2024-0008.json
