+++ pkg/extract/nvd/feed/cve/v2/testdata/golden/semver-range/datasource.json
@@ -1,0 +1,4 @@
+{
+	"id": "nvd-feed-cve-v2",
+	"name": "NVD Feed v2 CVE"
+}
