+++ pkg/extract/nvd/feed/cve/v2/v2.go
@@ -1,19 +1,72 @@
+// Package v2 extracts NVD Feed v2 CVE data with cpematch expansion.
+//
+// Key differences from the api/cve extractor:
+//   - Index keys use part:vendor:product:version format (ANY for wildcard).
+//   - Ranged CPE matches keep their range criterion and additionally expand
+//     the cpematch feed. Unknown (non-semver) ranges add every expanded
+//     version as an exact-match criterion; SEMVER ranges add only the
+//     versions the range criterion cannot cover — non-semver versions and
+//     the rare semver version that falls outside the range (Strategy E, see
+//     docs/nvd-extract-strategy.md).
+//   - cpematch files are located via a matchCriteriaId → path index built
+//     once by walking the cpematch feed directory.
 package v2
 
 import (
+	"context"
+	"fmt"
 	"io/fs"
-	"log/slog"
+	"log"
 	"path/filepath"
+	"runtime"
+	"slices"
+	"strings"
 
+	"github.com/hashicorp/go-version"
+	"github.com/knqyf263/go-cpe/common"
+	"github.com/knqyf263/go-cpe/naming"
 	"github.com/pkg/errors"
+	"golang.org/x/sync/errgroup"
 
+	dataTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data"
+	cweTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/cwe"
+	detectionType "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection"
+	conditionTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition"
+	criteriaTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria"
+	criterionTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion"
+	vcTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion"
+	affectedTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/affected"
+	rangeTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/affected/range"
+	fixstatusTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/fixstatus"
+	criterionpackageTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/package"
+	cpePackageTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/condition/criteria/criterion/versioncriterion/package/cpe"
+	segmentTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/segment"
+	ecosystemTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/detection/segment/ecosystem"
+	referenceTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/reference"
+	severityTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/severity"
+	v2Types "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/severity/cvss/v2"
+	v30Types "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/severity/cvss/v30"
+	v31Types "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/severity/cvss/v31"
+	v40Types "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/severity/cvss/v40"
+	vulnerabilityTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/vulnerability"
+	vulnerabilityContentTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/data/vulnerability/content"
+	datasourceTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/datasource"
+	repositoryTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/datasource/repository"
+	sourceTypes "github.com/MaineK00n/vuls-data-update/pkg/extract/types/source"
 	"github.com/MaineK00n/vuls-data-update/pkg/extract/util"
+	utilgit "github.com/MaineK00n/vuls-data-update/pkg/extract/util/git"
+	utiljson "github.com/MaineK00n/vuls-data-update/pkg/extract/util/json"
+	utiltime "github.com/MaineK00n/vuls-data-update/pkg/extract/util/time"
+	cpematchTypes "github.com/MaineK00n/vuls-data-update/pkg/fetch/nvd/feed/cpematch/v2"
+	cveTypes "github.com/MaineK00n/vuls-data-update/pkg/fetch/nvd/feed/cve/v2"
 )
 
 type options struct {
-	dir string
+	dir         string
+	concurrency int
 }
 
+// Option configures the extraction.
 type Option interface {
 	apply(*options)
 }
@@ -24,13 +77,35 @@ func (d dirOption) apply(opts *options) {
 	opts.dir = string(d)
 }
 
+// WithDir sets the output directory.
 func WithDir(dir string) Option {
 	return dirOption(dir)
 }
 
-func Extract(args string, opts ...Option) error {
+type concurrencyOption int
+
+func (c concurrencyOption) apply(opts *options) {
+	opts.concurrency = int(c)
+}
+
+// WithConcurrency sets the number of parallel workers.
+func WithConcurrency(concurrency int) Option {
+	return concurrencyOption(concurrency)
+}
+
+type extractor struct {
+	cpematchDir   string
+	cpematchIndex map[string]string // matchCriteriaId → cpematch file path
+	outputDir     string
+	r             *utiljson.JSONReader
+}
+
+// Extract processes NVD Feed v2 CVE data, expanding cpematch entries
+// and producing extracted detection data.
+func Extract(cveDir, cpematchDir string, opts ...Option) error {
 	options := &options{
-		dir: filepath.Join(util.CacheDir(), "extract", "", ""),
+		dir:         filepath.Join(util.CacheDir(), "extract", "nvd", "feed-v2", "cve"),
+		concurrency: runtime.NumCPU(),
 	}
 
 	for _, o := range opts {
@@ -41,27 +116,528 @@ func Extract(args string, opts ...Option) error {
 		return errors.Wrapf(err, "remove %s", options.dir)
 	}
 
-	slog.Info("Extract NVD Feed CVE v2")
-	if err := filepath.WalkDir(args, func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			return err
-		}
+	log.Printf("[INFO] Extract NVD Feed v2 CVE")
 
-		if d.IsDir() {
-			if d.Name() == ".git" {
-				return filepath.SkipDir
+	cpematchIndex, err := buildCpematchIndex(cpematchDir)
+	if err != nil {
+		return errors.Wrap(err, "build cpematch index")
+	}
+
+	g, ctx := errgroup.WithContext(context.Background())
+	g.SetLimit(options.concurrency)
+
+	reqChan := make(chan string)
+	g.Go(func() error {
+		defer close(reqChan)
+		if err := filepath.WalkDir(cveDir, func(path string, d fs.DirEntry, err error) error {
+			if err != nil {
+				return err
+			}
+
+			if d.IsDir() || filepath.Ext(path) != ".json" {
+				return nil
+			}
+
+			select {
+			case reqChan <- path:
+			case <-ctx.Done():
+				return ctx.Err()
 			}
 			return nil
+		}); err != nil {
+			return errors.Wrapf(err, "walk %s", cveDir)
 		}
-
-		if filepath.Ext(path) != ".json" {
-			return nil
-		}
-
 		return nil
-	}); err != nil {
-		return errors.Wrapf(err, "walk %s", args)
+	})
+
+	for i := 0; i < options.concurrency; i++ {
+		g.Go(func() error {
+			for path := range reqChan {
+				if err := extract(path, cveDir, cpematchDir, cpematchIndex, options.dir); err != nil {
+					return errors.Wrapf(err, "extract %s", path)
+				}
+			}
+			return nil
+		})
+	}
+
+	if err := g.Wait(); err != nil {
+		return errors.Wrapf(err, "wait for walk")
+	}
+
+	if err := util.Write(filepath.Join(options.dir, "datasource.json"), datasourceTypes.DataSource{
+		ID:   sourceTypes.NVDFeedCVEv2,
+		Name: new("NVD Feed v2 CVE"),
+		Raw: func() []repositoryTypes.Repository {
+			var res []repositoryTypes.Repository
+			cveGit, _ := utilgit.GetDataSourceRepository(cveDir)
+			if cveGit != nil {
+				res = append(res, *cveGit)
+			}
+			cpematchGit, _ := utilgit.GetDataSourceRepository(cpematchDir)
+			if cpematchGit != nil {
+				res = append(res, *cpematchGit)
+			}
+			return res
+		}(),
+		Extracted: func() *repositoryTypes.Repository {
+			if u, err := utilgit.GetOrigin(options.dir); err == nil {
+				return &repositoryTypes.Repository{
+					URL: u,
+				}
+			}
+			return nil
+		}(),
+	}, false); err != nil {
+		return errors.Wrapf(err, "write %s", filepath.Join(options.dir, "datasource.json"))
 	}
 
 	return nil
+}
+
+func extract(cvePath, cveDir, cpematchDir string, cpematchIndex map[string]string, outputDir string) error {
+	e := extractor{
+		cpematchDir:   cpematchDir,
+		cpematchIndex: cpematchIndex,
+		outputDir:     outputDir,
+		r:             utiljson.NewJSONReader(),
+	}
+
+	var fetched cveTypes.CVE
+	if err := e.r.Read(cvePath, cveDir, &fetched); err != nil {
+		return errors.Wrapf(err, "read json %s", cvePath)
+	}
+
+	data, err := e.buildData(fetched)
+	if err != nil {
+		return errors.Wrapf(err, "buildData %s", cvePath)
+	}
+
+	if err := util.Write(filepath.Join(e.outputDir, "data", filepath.Base(filepath.Dir(cvePath)), fmt.Sprintf("%s.json", data.ID)), data, true); err != nil {
+		return errors.Wrapf(err, "write %s", filepath.Join(e.outputDir, "data", filepath.Base(filepath.Dir(cvePath)), fmt.Sprintf("%s.json", data.ID)))
+	}
+	return nil
+}
+
+func (e extractor) buildData(fetched cveTypes.CVE) (dataTypes.Data, error) {
+	ds, err := func() ([]detectionType.Detection, error) {
+		switch len(fetched.Configurations) {
+		case 0:
+			return nil, nil
+		default:
+			rootCriteria := criteriaTypes.Criteria{
+				Operator:  criteriaTypes.CriteriaOperatorTypeOR,
+				Criterias: make([]criteriaTypes.Criteria, 0, len(fetched.Configurations)),
+			}
+			for _, c := range fetched.Configurations {
+				ca, err := e.configurationToCriteria(c)
+				if err != nil {
+					return nil, errors.Wrapf(err, "configuration to criteria. ID: %s", fetched.ID)
+				}
+				rootCriteria.Criterias = append(rootCriteria.Criterias, ca)
+			}
+			return []detectionType.Detection{{
+				Ecosystem: ecosystemTypes.EcosystemTypeCPE,
+				Conditions: []conditionTypes.Condition{{
+					Criteria: rootCriteria,
+				}},
+			}}, nil
+		}
+	}()
+	if err != nil {
+		return dataTypes.Data{}, errors.Wrapf(err, "build detection. ID: %s", fetched.ID)
+	}
+
+	ss := make([]severityTypes.Severity, 0, len(fetched.Metrics.CVSSMetricV2)+len(fetched.Metrics.CVSSMetricV30)+len(fetched.Metrics.CVSSMetricV31)+len(fetched.Metrics.CVSSMetricV40))
+	switch cap(ss) {
+	case 0:
+		ss = nil
+	default:
+		for _, c := range fetched.Metrics.CVSSMetricV2 {
+			sv2, err := v2Types.Parse(c.CvssData.VectorString)
+			if err != nil {
+				return dataTypes.Data{}, errors.Wrapf(err, "cvss v2 parse. vector: %s", c.CvssData.VectorString)
+			}
+			ss = append(ss, severityTypes.Severity{
+				Type:   severityTypes.SeverityTypeCVSSv2,
+				Source: c.Source,
+				CVSSv2: sv2,
+			})
+		}
+		for _, c := range fetched.Metrics.CVSSMetricV30 {
+			v30, err := v30Types.Parse(c.CVSSData.VectorString)
+			if err != nil {
+				return dataTypes.Data{}, errors.Wrapf(err, "cvss v30 parse. vector: %s", c.CVSSData.VectorString)
+			}
+			ss = append(ss, severityTypes.Severity{
+				Type:    severityTypes.SeverityTypeCVSSv30,
+				Source:  c.Source,
+				CVSSv30: v30,
+			})
+		}
+		for _, c := range fetched.Metrics.CVSSMetricV31 {
+			v31, err := v31Types.Parse(c.CVSSData.VectorString)
+			if err != nil {
+				return dataTypes.Data{}, errors.Wrapf(err, "cvss v31 parse. vector: %s", c.CVSSData.VectorString)
+			}
+			ss = append(ss, severityTypes.Severity{
+				Type:    severityTypes.SeverityTypeCVSSv31,
+				Source:  c.Source,
+				CVSSv31: v31,
+			})
+		}
+		for _, c := range fetched.Metrics.CVSSMetricV40 {
+			v40, err := v40Types.Parse(c.CVSSData.VectorString)
+			if err != nil {
+				return dataTypes.Data{}, errors.Wrapf(err, "cvss v40 parse. vector: %s", c.CVSSData.VectorString)
+			}
+			ss = append(ss, severityTypes.Severity{
+				Type:    severityTypes.SeverityTypeCVSSv40,
+				Source:  c.Source,
+				CVSSv40: v40,
+			})
+		}
+	}
+
+	return dataTypes.Data{
+		ID: dataTypes.RootID(fetched.ID),
+		Vulnerabilities: []vulnerabilityTypes.Vulnerability{
+			{
+				Content: vulnerabilityContentTypes.Content{
+					ID: vulnerabilityContentTypes.VulnerabilityID(fetched.ID),
+					Description: func() string {
+						for _, d := range fetched.Descriptions {
+							if d.Lang == "en" {
+								return d.Value
+							}
+						}
+						return ""
+					}(),
+					Severity: ss,
+					CWE: func() []cweTypes.CWE {
+						if len(fetched.Weaknesses) == 0 {
+							return nil
+						}
+
+						m := make(map[string][]string)
+						for _, w := range fetched.Weaknesses {
+							for _, d := range w.Description {
+								m[w.Source] = append(m[w.Source], d.Value)
+							}
+						}
+						cs := make([]cweTypes.CWE, 0, len(m))
+						for s, vs := range m {
+							cs = append(cs, cweTypes.CWE{Source: s, CWE: util.Unique(vs)})
+						}
+						return cs
+					}(),
+					References: func() []referenceTypes.Reference {
+						refs := make([]referenceTypes.Reference, 0, 1+len(fetched.References))
+						refs = append(refs, referenceTypes.Reference{
+							Source: "nvd.nist.gov",
+							URL:    fmt.Sprintf("https://nvd.nist.gov/vuln/detail/%s", fetched.ID),
+						})
+						for _, r := range fetched.References {
+							refs = append(refs, referenceTypes.Reference{
+								Source: r.Source,
+								URL:    r.URL,
+							})
+						}
+						return refs
+					}(),
+					Published: utiltime.Parse([]string{"2006-01-02T15:04:05.000"}, fetched.Published),
+					Modified:  utiltime.Parse([]string{"2006-01-02T15:04:05.000"}, fetched.LastModified),
+				},
+				Segments: []segmentTypes.Segment{{Ecosystem: ecosystemTypes.EcosystemTypeCPE}},
+			},
+		},
+		Detections: ds,
+		DataSource: sourceTypes.Source{
+			ID:   sourceTypes.NVDFeedCVEv2,
+			Raws: e.r.Paths(),
+		},
+	}, nil
+}
+
+func (e extractor) configurationToCriteria(config cveTypes.Config) (criteriaTypes.Criteria, error) {
+	if config.Negate {
+		return criteriaTypes.Criteria{}, errors.New("negate in Config is not implemented")
+	}
+
+	ca := criteriaTypes.Criteria{}
+	switch config.Operator {
+	case "AND":
+		ca.Operator = criteriaTypes.CriteriaOperatorTypeAND
+	case "OR", "":
+		ca.Operator = criteriaTypes.CriteriaOperatorTypeOR
+	default:
+		return criteriaTypes.Criteria{}, errors.Errorf("invalid configuration operator: %s", config.Operator)
+	}
+
+	ca.Criterias = make([]criteriaTypes.Criteria, 0, len(config.Nodes))
+	for _, n := range config.Nodes {
+		child, err := e.nodeToCriteria(n)
+		if err != nil {
+			return criteriaTypes.Criteria{}, errors.Wrap(err, "nodeToCriteria")
+		}
+		ca.Criterias = append(ca.Criterias, child)
+	}
+	return ca, nil
+}
+
+func (e extractor) nodeToCriteria(n cveTypes.Node) (criteriaTypes.Criteria, error) {
+	ca := criteriaTypes.Criteria{}
+	switch n.Operator {
+	case "AND":
+		ca.Operator = criteriaTypes.CriteriaOperatorTypeAND
+	case "OR":
+		ca.Operator = criteriaTypes.CriteriaOperatorTypeOR
+	default:
+		return criteriaTypes.Criteria{}, errors.Errorf("invalid node operator: %s", n.Operator)
+	}
+
+	ca.Criterias = make([]criteriaTypes.Criteria, 0, len(n.CPEMatch))
+	for _, match := range n.CPEMatch {
+		if _, err := naming.UnbindFS(match.Criteria); err != nil {
+			return criteriaTypes.Criteria{}, errors.Wrapf(err, "invalid format. CPE: %s", match.Criteria)
+		}
+
+		hasRange := match.VersionStartIncluding != "" || match.VersionStartExcluding != "" ||
+			match.VersionEndIncluding != "" || match.VersionEndExcluding != ""
+
+		rangeType := decideRangeType(match)
+
+		cn := criterionTypes.Criterion{
+			Type: criterionTypes.CriterionTypeVersion,
+			Version: &vcTypes.Criterion{
+				Vulnerable: match.Vulnerable,
+				FixStatus: func() *fixstatusTypes.FixStatus {
+					if match.Vulnerable {
+						return &fixstatusTypes.FixStatus{Class: fixstatusTypes.ClassUnknown}
+					}
+					return nil
+				}(),
+				Package: criterionpackageTypes.Package{
+					Type: criterionpackageTypes.PackageTypeCPE,
+					CPE:  new(cpePackageTypes.CPE(match.Criteria)),
+				},
+				Affected: func() *affectedTypes.Affected {
+					if !hasRange {
+						return nil
+					}
+					return &affectedTypes.Affected{
+						Type: rangeType,
+						Range: []rangeTypes.Range{{
+							GreaterEqual: match.VersionStartIncluding,
+							GreaterThan:  match.VersionStartExcluding,
+							LessEqual:    match.VersionEndIncluding,
+							LessThan:     match.VersionEndExcluding,
+						}},
+					}
+				}(),
+			},
+		}
+
+		cns := []criterionTypes.Criterion{cn}
+
+		// Strategy E: a ranged match keeps its range criterion above and, in
+		// addition, expands the cpematch feed into exact-match criteria.
+		//   - Unknown range: the range cannot be evaluated at detection time,
+		//     so every expanded version is added.
+		//   - SEMVER range: the range criterion already covers every
+		//     semver-parseable version inside the range; only the versions it
+		//     cannot cover (non-semver, or the rare semver version outside the
+		//     range) need an explicit exact-match criterion.
+		// See docs/nvd-extract-strategy.md.
+		if hasRange {
+			ns, err := e.cpeNamesFromCpematch(match.MatchCriteriaID)
+			if err != nil {
+				log.Printf("[WARN] cpematch lookup failed for %s (criteria=%s): %v", match.MatchCriteriaID, match.Criteria, err)
+			} else {
+				cns = slices.Grow(cns, len(ns))
+				for _, n := range ns {
+					if rangeType == rangeTypes.RangeTypeSEMVER && coveredBySemverRange(n, match) {
+						continue
+					}
+					cns = append(cns, exactCriterion(n, match.Vulnerable))
+				}
+			}
+		}
+
+		ca.Criterias = append(ca.Criterias, criteriaTypes.Criteria{
+			Operator:   criteriaTypes.CriteriaOperatorTypeOR,
+			Criterions: cns,
+		})
+	}
+	return ca, nil
+}
+
+// exactCriterion builds an exact-match version criterion for a single
+// expanded cpematch CPE name.
+func exactCriterion(cpeName string, vulnerable bool) criterionTypes.Criterion {
+	return criterionTypes.Criterion{
+		Type: criterionTypes.CriterionTypeVersion,
+		Version: &vcTypes.Criterion{
+			Vulnerable: vulnerable,
+			FixStatus: func() *fixstatusTypes.FixStatus {
+				if vulnerable {
+					return &fixstatusTypes.FixStatus{Class: fixstatusTypes.ClassUnknown}
+				}
+				return nil
+			}(),
+			Package: criterionpackageTypes.Package{
+				Type: criterionpackageTypes.PackageTypeCPE,
+				CPE:  new(cpePackageTypes.CPE(cpeName)),
+			},
+		},
+	}
+}
+
+// coveredBySemverRange reports whether the expanded cpematch CPE name n is
+// already covered by a SEMVER range criterion — i.e. it carries a concrete,
+// semver-parseable version that lies within match's range. When it returns
+// false, n needs an explicit exact-match criterion because the range
+// criterion cannot match it: either the version is non-semver, or it is the
+// rare semver version that NVD lists yet which falls outside the declared
+// range (~0.017% of cases — segment-count or pre-release ordering quirks in
+// NVD data; see docs/nvd-extract-verification.md V5).
+func coveredBySemverRange(n string, match cveTypes.CPEMatch) bool {
+	wfn, err := naming.UnbindFS(n)
+	if err != nil {
+		return false
+	}
+	// GetString returns WFN-escaped form; unescape for version parsing.
+	ver := unescapeWFN(wfn.GetString(common.AttributeVersion))
+	if ver == "" || ver == "*" || ver == "-" {
+		return true // not a concrete version; nothing to add
+	}
+	sv, err := version.NewSemver(ver)
+	if err != nil {
+		return false
+	}
+	return versionInRange(sv, match)
+}
+
+// versionInRange checks if the given semver falls within the range.
+func versionInRange(v *version.Version, match cveTypes.CPEMatch) bool {
+	if match.VersionStartIncluding != "" {
+		bound, err := version.NewSemver(match.VersionStartIncluding)
+		if err != nil {
+			return false
+		}
+		if v.LessThan(bound) {
+			return false
+		}
+	}
+	if match.VersionStartExcluding != "" {
+		bound, err := version.NewSemver(match.VersionStartExcluding)
+		if err != nil {
+			return false
+		}
+		if !v.GreaterThan(bound) {
+			return false
+		}
+	}
+	if match.VersionEndIncluding != "" {
+		bound, err := version.NewSemver(match.VersionEndIncluding)
+		if err != nil {
+			return false
+		}
+		if v.GreaterThan(bound) {
+			return false
+		}
+	}
+	if match.VersionEndExcluding != "" {
+		bound, err := version.NewSemver(match.VersionEndExcluding)
+		if err != nil {
+			return false
+		}
+		if !v.LessThan(bound) {
+			return false
+		}
+	}
+	return true
+}
+
+// buildCpematchIndex walks the cpematch feed directory and maps each
+// matchCriteriaId (the JSON file's basename) to its file path. NVD stores
+// cpematch files in FNV-hashed subdirectories keyed by vendor:product, but
+// the CVE feed and the cpematch feed disagree on some product spellings
+// (e.g. nx-os vs nx_os), so that hash is not reproducible from the CVE side.
+// Indexing by matchCriteriaId sidesteps the problem entirely.
+func buildCpematchIndex(cpematchDir string) (map[string]string, error) {
+	index := make(map[string]string)
+	if err := filepath.WalkDir(cpematchDir, func(path string, d fs.DirEntry, err error) error {
+		if err != nil {
+			return err
+		}
+		if d.IsDir() || filepath.Ext(path) != ".json" {
+			return nil
+		}
+		index[strings.TrimSuffix(d.Name(), ".json")] = path
+		return nil
+	}); err != nil {
+		return nil, errors.Wrapf(err, "walk %s", cpematchDir)
+	}
+	return index, nil
+}
+
+// cpeNamesFromCpematch loads the cpematch file for the given matchCriteriaId
+// via the prebuilt index and returns its expanded CPE names.
+func (e extractor) cpeNamesFromCpematch(matchCriteriaId string) ([]string, error) {
+	path, ok := e.cpematchIndex[matchCriteriaId]
+	if !ok {
+		return nil, errors.Errorf("cpematch not found for matchCriteriaId=%s", matchCriteriaId)
+	}
+
+	var cpeMatch cpematchTypes.MatchCriteria
+	if err := e.r.Read(path, e.cpematchDir, &cpeMatch); err != nil {
+		return nil, errors.Wrapf(err, "read json %s", path)
+	}
+
+	ns := make([]string, 0, len(cpeMatch.Matches))
+	for _, m := range cpeMatch.Matches {
+		ns = append(ns, m.CPEName)
+	}
+	return ns, nil
+}
+
+// decideRangeType checks if all range endpoints are valid semver.
+func decideRangeType(match cveTypes.CPEMatch) rangeTypes.RangeType {
+	for _, v := range []string{match.VersionStartIncluding, match.VersionStartExcluding, match.VersionEndIncluding, match.VersionEndExcluding} {
+		if v == "" {
+			continue
+		}
+		if _, err := version.NewSemver(v); err != nil {
+			return rangeTypes.RangeTypeUnknown
+		}
+	}
+	return rangeTypes.RangeTypeSEMVER
+}
+
+// IndexKey returns the index key for a CPE formatted string
+// in the format "part:vendor:product:version".
+// For version=* (ANY), returns "ANY".
+func IndexKey(cpe string) (string, error) {
+	wfn, err := naming.UnbindFS(cpe)
+	if err != nil {
+		return "", errors.Wrapf(err, "unbind CPE %q", cpe)
+	}
+
+	part := wfn.GetString(common.AttributePart)
+	vendor := unescapeWFN(wfn.GetString(common.AttributeVendor))
+	product := unescapeWFN(wfn.GetString(common.AttributeProduct))
+	ver := unescapeWFN(wfn.GetString(common.AttributeVersion))
+
+	if ver == "" || ver == "*" {
+		ver = "ANY"
+	}
+
+	return fmt.Sprintf("%s:%s:%s:%s", part, vendor, product, ver), nil
+}
+
+// unescapeWFN removes WFN backslash escaping from attribute values.
+// e.g. "7\\.1\\.2" → "7.1.2"
+func unescapeWFN(s string) string {
+	return strings.ReplaceAll(s, "\\", "")
 }
