+++ pkg/extract/nvd/feed/cve/v2/v2_test.go
@@ -10,38 +10,61 @@ import (
 
 func TestExtract(t *testing.T) {
 	tests := []struct {
-		name     string
-		args     string
-		hasError bool
+		name        string
+		cveDir      string
+		cpematchDir string
+		golden      string
+		hasError    bool
 	}{
 		{
-			name: "happy",
-			args: "./testdata/fixtures",
+			name:        "semver-range",
+			cveDir:      "./testdata/fixtures/semver-range/vuls-data-raw-nvd-feed-cve-v2",
+			cpematchDir: "./testdata/fixtures/semver-range/vuls-data-raw-nvd-feed-cpematch-v2",
+			golden:      "./testdata/golden/semver-range",
+		},
+		{
+			// SEMVER range whose cpematch mixes semver and non-semver versions:
+			// Strategy E adds only the non-semver versions as exact-match
+			// criteria; semver versions inside the range stay covered by the
+			// range criterion alone.
+			name:        "semver-range-mixed",
+			cveDir:      "./testdata/fixtures/semver-range-mixed/vuls-data-raw-nvd-feed-cve-v2",
+			cpematchDir: "./testdata/fixtures/semver-range-mixed/vuls-data-raw-nvd-feed-cpematch-v2",
+			golden:      "./testdata/golden/semver-range-mixed",
+		},
+		{
+			name:        "non-semver-range",
+			cveDir:      "./testdata/fixtures/non-semver-range/vuls-data-raw-nvd-feed-cve-v2",
+			cpematchDir: "./testdata/fixtures/non-semver-range/vuls-data-raw-nvd-feed-cpematch-v2",
+			golden:      "./testdata/golden/non-semver-range",
+		},
+		{
+			name:        "exact-match",
+			cveDir:      "./testdata/fixtures/exact-match/vuls-data-raw-nvd-feed-cve-v2",
+			cpematchDir: "./testdata/fixtures/exact-match/vuls-data-raw-nvd-feed-cpematch-v2",
+			golden:      "./testdata/golden/exact-match",
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			dir := t.TempDir()
-			err := v2.Extract(tt.args, v2.WithDir(dir))
+			err := v2.Extract(tt.cveDir, tt.cpematchDir, v2.WithDir(dir))
 			switch {
 			case err != nil && !tt.hasError:
 				t.Error("unexpected error:", err)
 			case err == nil && tt.hasError:
 				t.Error("expected error has not occurred")
-			case err != nil && tt.hasError:
-				// error was expected and occurred, test passed
-				return
-			default:
-				ep, err := filepath.Abs(filepath.Join("testdata", "golden"))
-				if err != nil {
-					t.Error("unexpected error:", err)
-				}
-				gp, err := filepath.Abs(dir)
-				if err != nil {
-					t.Error("unexpected error:", err)
-				}
-				utiltest.Diff(t, ep, gp)
 			}
+
+			ep, err := filepath.Abs(tt.golden)
+			if err != nil {
+				t.Error("unexpected error:", err)
+			}
+			gp, err := filepath.Abs(dir)
+			if err != nil {
+				t.Error("unexpected error:", err)
+			}
+			utiltest.Diff(t, ep, gp)
 		})
 	}
 }
